Debugged

+++ 02-install.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -ex
+
 sudo -v
 
 # MATE APPEARANCE
@@ -8,13 +10,17 @@ dconf write /org/mate/marco/general/center-new-window true
 
 # GIT
 rm ~/.gitconfig
-ln -s ~/dotfiles/git.conf ~/.gitconfig
+ln -s ~/dotfiles/git/git.conf ~/.gitconfig
 
 # Terminator
 sudo pacman -S terminator
 mkdir -p ~/.config/terminator/
-ln -s ~/dotfiles/terminator/terminator.conf ~/.config/terminator/config
-ln -s ~/dotfiles/terminator/terminator.desktop /etc/xdg/autostart/terminator.desktop
+if [ ! -e $HOME/.config/terminator/config ]; then
+    ln -s ~/dotfiles/terminator/terminator.conf ~/.config/terminator/config
+fi
+if [ ! -e /etc/xdg/autostart/terminator.desktop ]; then
+    sudo ln -s ~/dotfiles/terminator/terminator.desktop /etc/xdg/autostart/terminator.desktop
+fi
 
 # SUBLIME TEXT (editor, keymap, config, packages)
 sudo pacman -S gtk2 libpng
@@ -36,24 +42,35 @@ sudo gpasswd -a $USER docker
 # PHP
 sudo pacman -S php php-intl
 sudo rm /etc/php/php.ini
-sudo ln -s ~/dotfiles/php/php.ini /etc/php/php.ini
+if [ ! -e /etc/php/php.ini ]; then
+    sudo ln -s ~/dotfiles/php/php.ini /etc/php/php.ini
+fi
 
 # PHP TOOLS
-curl -O http://getcomposer.org/composer.phar
-sudo mv composer.phar /usr/local/bin/composer
+if [ ! -e /usr/local/bin/composer ]; then
+    curl -O http://getcomposer.org/composer.phar
+    sudo mv composer.phar /usr/local/bin/composer
+fi
 sudo chmod a+x /usr/local/bin/composer
 composer global require "squizlabs/php_codesniffer=~2.3"
 composer global require "phpmd/phpmd=~2.2"
 composer global require "phpunit/phpunit=~4.6"
 
 # Z
-git clone git@github.com:rupa/z.git ~/.rupaz
+if [ ! -d ~/.rupaz ]; then
+    git clone git@github.com:rupa/z.git ~/.rupaz
+fi
 
 # ZSH
 sudo pacman -S zsh zsh-completions
-git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
-cp ~/dotfiles/zsh/zshrc ~/.zshrc
-ln -s ~/dotfiles/zsh/owlycode.zsh-theme ~/.oh-my-zsh/themes/owlycode.zsh-theme
+if [ ! -d ~/.oh-my-zsh ]; then
+	git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
+fi
+if [ ! -e $HOME/.zshrc ]; then
+    ln -s ~/dotfiles/zsh/zshrc ~/.zshrc
+fi
+if [ ! -e $HOME/.oh-my-zsh/themes/owlycode.zsh-theme ]; then
+	ln -s ~/dotfiles/zsh/owlycode.zsh-theme ~/.oh-my-zsh/themes/owlycode.zsh-theme
+fi
 chsh -s /bin/zsh
-
